// src/modules/configuration/model/MemberWorkGroup.ts
export interface MemberWorkGroup  {
    IDWorkGroupMember: number;
    AccountID: number;
    IDLnFunctionalID: number;
    State: number;
    StateName: number;
    UptoDate: Date;
    Since: Date;
    IDAccount: number;
    IDOffice: number;
    AccountName: string;
    FunctionalIDName: string;
    OfficeName: string;
}